requirements-dev.txt
-r requirements.txt
pip>=24.1.2,<25.3
pytest-homeassistant-custom-component==0.13.224
pre-commit
isort
black
codespell
mypy
colorlog